bash: Refactor prompt and prompt colors
Some of the previous prompt colors were approximations of colors in the
solarized theme from the 256 default terminal colors.  While this
approach works okay (and is actually required if not using a solarized
terminal color palette), using colors from the user-defined terminal
color palette gives the exact theme colors and thus is preferred.